feat(CarlaNet): comprehensive feature-complete implementation pass
Protocol & serialization:
- NullableDoubleFormatter: correct std::optional<T> encoding [false]/[true,val]
  (was using nil/raw — now matches carla/MsgPackAdaptors.h exactly)
- CommandFormatter: serialize all 22 Command variants as [[idx, payload]]
  per Command::MSGPACK_DEFINE_ARRAY(command) + std::variant adaptor
- CommandResponseFormatter: deserialize Response<ActorId> wire format
  ([[0,["err"]]] or [[1,actor_id]]) — CommandResponse.h is a type alias
- DebugShapeFormatter: serialize all 5 DebugShape primitives as [[idx,payload]]
- UnpackResult<T>: detect ResponseError (variant idx=0) and throw CarlaRpcException
- Reader loop: copy ReadOnlySequence to flat array (MessagePackStreamReader
  reclaims internal buffers on next ReadAsync — large responses were invalid)

Type fixes (source-verified against C++ headers):
- TrafficLightState, VehicleFailureState, AttachmentType, VehicleDoor: uint→byte
  (all are : uint8_t in source; ActorState was already correct)
- Actor.SemanticTags, Actor.StreamToken: IReadOnlyList<byte>/RawToken→byte[]
  (both are std::vector<unsigned char> — raw binary, not MSGPACK_DEFINE_ARRAY)
- RawToken: kept for EpisodeInfo.Token (still uses streaming::Token with DEFINE_ARRAY)
- CommandResponse: replaced flat [MessagePackObject] struct with proper type
  backed by CommandResponseFormatter (was wrong; source: Response<ActorId>)
- SensorFrame.PayloadBytes: added byte[] accessor for Python interop

Transport:
- StreamToken.Parse: DNS resolution for hostname server hosts ("localhost")
- CarlaClient.StartWorldObserverAsync(): subscribes to episode state stream
  (FWorldObserver / EpisodeStateSerializer) and caches actor snapshots inline
  without CarlaNet.Sensors dependency
- CarlaClient: GetActorTransform, GetActorVelocity, GetActorAngularVelocity,
  GetActorAcceleration, GetVehicleControl, GetActorSnapshot, GetCachedActorIds
  — all read from the in-memory world observer cache
- CarlaClient.SubscribeToGBufferAsync: get_gbuffer_token RPC + stream sub
- CarlaClient.DrawDebugShapeAsync: was missing, added
- CarlaClient.SpawnVehicleAsync, SpawnCameraAsync: convenience helpers
- ActorSnapshot: public sealed class for per-actor cached state
- GlobalUsings: added System.Linq for LINQ methods in CarlaClient

Python:
- carlanet.py: comprehensive Python compatibility layer (§12)
  - Client: get_world, apply_batch, apply_batch_sync, recorder, replayer
  - World: get_map, get_blueprint_library, spawn_actor, try_spawn_actor,
    get_actors (uses world observer cache), tick, wait_for_tick, weather
  - Map: wraps MapInfo, get_spawn_points()
  - BlueprintLibrary: filter(glob), find(id), iter/len/getitem
  - ActorBlueprint: has_attribute, get_attribute, set_attribute, to_description
  - Actor wrapper: get_transform, get_velocity, get_control, apply_control,
    set_autopilot, set_light_state, physics, listen/stop, destroy
  - command.*: SpawnActor+.then(), DestroyActor, SetAutopilot, ApplyVehicleControl,
    ApplyTransform, ApplyLocation, FutureActor, Response
  - TrafficManager: full 20+ method coverage per TrafficManagerServer.h bindings
  - carlanet.py copied to publish/ for direct import
- carlanet_manual_control.py: raw CarlaNet pygame driver (pre-carlanet.py shim)
- carlanet_probe.py: end-to-end live server probe

All builds: 0 errors, 0 warnings.
Smoke test: offline checks pass (33/42 bytes — correct new optional encoding).

+++ CarlaNet/src/CarlaNet.Types/Rpc/Enums.cs
@@ -2,29 +2,38 @@
 //          TrafficLightState.h, VehicleDoor.h, VehicleWheels.h,
 //          VehicleFailureState.h, MapLayer.h, ObjectLabel.h, MaterialParameter.h,
 //          LightState.h, VehicleLightState.h
+// Underlying types verified against source headers (uint8_t vs int).
 namespace CarlaNet.Types.Rpc.Enums;
 
+// ActorAttributeType has no explicit underlying type in C++ (defaults to int)
 public enum ActorAttributeType : uint
 { Bool, Int, Float, String, RGBColor, Vector, SIZE, INVALID }
 
+// ActorState : uint8_t (carla/rpc/ActorState.h)
 public enum ActorState : byte
 { Invalid, Active, Dormant, PendingKill }
 
-public enum AttachmentType : uint
+// AttachmentType : uint8_t (carla/rpc/AttachmentType.h)
+public enum AttachmentType : byte
 { Rigid, SpringArm, SpringArmGhost, SIZE, INVALID }
 
-public enum TrafficLightState : uint
+// TrafficLightState : uint8_t (carla/rpc/TrafficLightState.h)
+public enum TrafficLightState : byte
 { Red, Yellow, Green, Off, Unknown, SIZE }
 
-public enum VehicleDoor : uint
+// VehicleDoor : uint8_t (carla/rpc/VehicleDoor.h)
+public enum VehicleDoor : byte
 { FL = 0, FR = 1, RL = 2, RR = 3, Hood = 4, Trunk = 5, All = 6 }
 
+// VehicleWheelLocation has no explicit underlying type in source
 public enum VehicleWheelLocation : uint
 { FL = 0, FR = 1, BL = 2, BR = 3, FrontWheel = 0, BackWheel = 1 }
 
-public enum VehicleFailureState : uint
+// VehicleFailureState : uint8_t (carla/rpc/VehicleFailureState.h)
+public enum VehicleFailureState : byte
 { None, Rollover, Engine, TirePuncture }
 
+// MapLayer : uint16_t (carla/rpc/MapLayer.h — using MapLayerType = uint16_t)
 [Flags]
 public enum MapLayer : ushort
 {
@@ -33,6 +42,7 @@ public enum MapLayer : ushort
     Walls = 0x100, All = 0xFFFF
 }
 
+// CityObjectLabel / ObjectLabel — byte in sensor data
 public enum CityObjectLabel : byte
 {
     None = 0, Roads = 1, Sidewalks = 2, Buildings = 3, Walls = 4, Fences = 5,
@@ -43,12 +53,15 @@ public enum CityObjectLabel : byte
     RailTrack = 27, GuardRail = 28, Rock = 29, Any = 0xFF
 }
 
+// MaterialParameter has no explicit underlying type in C++ (defaults to int)
 public enum MaterialParameter : uint
 { TexNormal, TexAoRoughnessMetallicEmissive, TexDiffuse, TexEmissive }
 
+// LightGroup : uint8_t (carla/rpc/LightState.h)
 public enum LightGroup : byte
 { None = 0, Vehicle = 1, Street = 2, Building = 3, Other = 4 }
 
+// VehicleLightState.flag_type = uint32_t (carla/rpc/VehicleLightState.h)
 [Flags]
 public enum VehicleLightStateFlags : uint
 {
